+++ cedar-policy-core/src/ast/partial_value.rs
@@ -81,77 +81,3 @@ impl std::fmt::Display for PartialValue {
         }
     }
 }
-
-// PANIC SAFETY: Unit Test Code
-#[allow(clippy::panic)]
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::evaluator::split;
-    use itertools::Either;
-
-    #[test]
-    fn split_values() {
-        let vs = [
-            PartialValue::Value(Value::from(1)),
-            PartialValue::Value(Value::from(2)),
-        ];
-        match split(vs) {
-            Either::Right(_) => panic!("expected values, got residuals"),
-            Either::Left(vs) => {
-                assert_eq!(vs.collect::<Vec<_>>(), vec![Value::from(1), Value::from(2)])
-            }
-        };
-    }
-
-    #[test]
-    fn split_residuals() {
-        let rs = [
-            PartialValue::Value(Value::from(1)),
-            PartialValue::Residual(Expr::val(2)),
-            PartialValue::Value(Value::from(3)),
-            PartialValue::Residual(Expr::val(4)),
-        ];
-        let expected = vec![Expr::val(1), Expr::val(2), Expr::val(3), Expr::val(4)];
-        match split(rs) {
-            Either::Left(_) => panic!("expected residuals, got values"),
-            Either::Right(rs) => {
-                assert_eq!(rs.collect::<Vec<_>>(), expected);
-            }
-        };
-    }
-
-    #[test]
-    fn split_residuals2() {
-        let rs = [
-            PartialValue::Value(Value::from(1)),
-            PartialValue::Value(Value::from(2)),
-            PartialValue::Residual(Expr::val(3)),
-            PartialValue::Residual(Expr::val(4)),
-        ];
-        let expected = vec![Expr::val(1), Expr::val(2), Expr::val(3), Expr::val(4)];
-        match split(rs) {
-            Either::Left(_) => panic!("expected residuals, got values"),
-            Either::Right(rs) => {
-                assert_eq!(rs.collect::<Vec<_>>(), expected);
-            }
-        };
-    }
-
-    #[test]
-    fn split_residuals3() {
-        let rs = [
-            PartialValue::Residual(Expr::val(1)),
-            PartialValue::Residual(Expr::val(2)),
-            PartialValue::Value(Value::from(3)),
-            PartialValue::Value(Value::from(4)),
-        ];
-        let expected = vec![Expr::val(1), Expr::val(2), Expr::val(3), Expr::val(4)];
-        match split(rs) {
-            Either::Left(_) => panic!("expected residuals, got values"),
-            Either::Right(rs) => {
-                assert_eq!(rs.collect::<Vec<_>>(), expected);
-            }
-        };
-    }
-}
